+++ ginga/ImageView.py
@@ -179,11 +179,24 @@ class ImageViewBase(Callback.Callbacks):
         self.t_.get_setting('limits').add_callback('set', self._set_limits_cb)
 
         # embedded image "profiles"
-        self.t_.add_defaults(profile_use_scale=False, profile_use_pan=False,
+        self.use_image_profile = False
+        self.t_.add_defaults(profile_use_scale=False,
+                             profile_use_pan=False,
                              profile_use_cuts=False,
                              profile_use_transform=False,
                              profile_use_rotation=False,
+                             profile_use_distribution=False,
+                             profile_use_contrast=False,
                              profile_use_color_map=False)
+        self.profile_keylist = ['flip_x', 'flip_y', 'swap_xy', 'scale',
+                                'pan', 'pan_coord', 'rot_deg', 'cuts',
+                                'color_algorithm',
+                                'color_map', 'intensity_map',
+                                'color_array', 'shift_array']
+        #for name in self.t_.keys():
+        for name in self.profile_keylist:
+            self.t_.get_setting(name).add_callback('set',
+                                                   self._update_profile_cb)
 
         # ICC profile support
         d = dict(icc_output_profile=None, icc_output_intent='perceptual',
@@ -194,15 +207,17 @@ class ImageViewBase(Callback.Callbacks):
             self.t_.get_setting(key).add_callback('set', self.icc_profile_cb)
 
         # viewer profile support
-        self.use_image_profile = False
-        self.profile_keylist = ['flip_x', 'flip_y', 'swap_xy', 'scale',
-                                'pan', 'pan_coord', 'rot_deg', 'cuts',
-                                'color_algorithm', 'color_hashsize',
-                                'color_map', 'intensity_map',
-                                'color_array', 'shift_array']
-        for name in self.t_.keys():
-            self.t_.get_setting(name).add_callback('set',
-                                                   self._update_profile_cb)
+        self.default_viewer_profile = None
+        self.t_.add_defaults(viewer_restore_scale=False,
+                             viewer_restore_pan=False,
+                             viewer_restore_cuts=False,
+                             viewer_restore_transform=False,
+                             viewer_restore_rotation=False,
+                             viewer_restore_distribution=False,
+                             viewer_restore_contrast=False,
+                             viewer_restore_color_map=False)
+
+        self.capture_default_viewer_profile()
 
         # Object that calculates auto cut levels
         name = self.t_.get('autocut_method', 'zscale')
@@ -828,60 +843,118 @@ class ImageViewBase(Callback.Callbacks):
         is loaded.  Either the embedded profile settings or the default
         settings are applied as specified in the channel preferences.
         """
+        # 1. copy the current viewer settings
+        tmpprof = Settings.SettingGroup()
+        self.t_.copy_settings(tmpprof, include_callbacks=False,
+                              keylist=self.profile_keylist,
+                              callback=False)
+
+        # 2. reset selected items in the copy to default profile
+        # if there is one
+        keylist1 = set([])
+        if self.default_viewer_profile is not None:
+            dvp = self.default_viewer_profile
+            if self.t_['viewer_restore_transform'] and 'flip_x' in dvp:
+                keylist1.update({'flip_x', 'flip_y', 'swap_xy'})
+
+            if self.t_['viewer_restore_scale'] and 'scale' in dvp:
+                keylist1.add('scale')
+
+            if self.t_['viewer_restore_pan'] and 'pan' in dvp:
+                keylist1.add('pan')
+
+            if self.t_['viewer_restore_rotation'] and 'rot_deg' in dvp:
+                keylist1.add('rot_deg')
+
+            if self.t_['viewer_restore_cuts'] and 'cuts' in dvp:
+                keylist1.add('cuts')
+
+            if self.t_['viewer_restore_distribution'] and 'color_algorithm' in dvp:
+                keylist1.add('color_algorithm')
+
+            if self.t_['viewer_restore_color_map'] and 'color_array' in dvp:
+                keylist1.update({'color_array', 'intensity_map'})
+
+            if self.t_['viewer_restore_contrast'] and 'shift_array' in dvp:
+                keylist1.add('shift_array')
+
+            dvp.copy_settings(tmpprof, keylist=list(keylist1), callback=False)
+
+        # 3. apply image-embedded profile selected items to the copy
+        keylist2 = set([])
         profile = image.get('profile', None)
-        keylist = []
+        if profile is not None:
+            if self.t_['profile_use_transform'] and 'flip_x' in profile:
+                keylist2.update({'flip_x', 'flip_y', 'swap_xy'})
+
+            if self.t_['profile_use_scale'] and 'scale' in profile:
+                keylist2.add('scale')
+
+            if self.t_['profile_use_pan'] and 'pan' in profile:
+                keylist2.add('pan')
+
+            if self.t_['profile_use_rotation'] and 'rot_deg' in profile:
+                keylist2.add('rot_deg')
+
+            if self.t_['profile_use_cuts'] and 'cuts' in profile:
+                keylist2.add('cuts')
+
+            if self.t_['profile_use_distribution'] and 'color_algorithm' in profile:
+                keylist2.add('color_algorithm')
+
+            if self.t_['profile_use_color_map'] and 'color_array' in profile:
+                keylist2.update({'color_array', 'intensity_map'})
+
+            if self.t_['profile_use_contrast'] and 'shift_array' in profile:
+                keylist2.add('shift_array')
+
+            profile.copy_settings(tmpprof, keylist=list(keylist2),
+                                  callback=False)
+
         with self.suppress_redraw:
-            if profile is not None:
-                if self.t_['profile_use_transform'] and 'flip_x' in profile:
-                    keylist.extend(['flip_x', 'flip_y', 'swap_xy'])
+            # NOTE: special hack for modified color maps:
+            # copy over the "color_map" setting but don't invoke callback
+            # because that would override color array, which may be modified
+            # by inversion, rotation, etc.
+            if 'color_array' in keylist2:
+                profile.copy_settings(self.t_, keylist=['color_map'],
+                                      callback=False)
+            elif 'color_array' in keylist1:
+                dvp.copy_settings(self.t_, keylist=['color_map'],
+                                  callback=False)
 
-                if self.t_['profile_use_scale'] and 'scale' in profile:
-                    keylist.extend(['scale'])
+            # 4. update our settings from the copy
+            keylist = list(keylist1.union(keylist2))
+            self.apply_profile(tmpprof, keylist=keylist)
 
-                if self.t_['profile_use_pan'] and 'pan' in profile:
-                    keylist.extend(['pan'])
-
-                if self.t_['profile_use_rotation'] and 'rot_deg' in profile:
-                    keylist.extend(['rot_deg'])
-
-                if self.t_['profile_use_cuts'] and 'cuts' in profile:
-                    keylist.extend(['cuts'])
-
-                if self.t_['profile_use_color_map'] and 'color_map' in profile:
-                    keylist.extend(['color_algorithm', 'color_hashsize',
-                                    'color_map', 'intensity_map',
-                                    'color_array', 'shift_array'])
-
-                self.apply_profile(profile, keylist=keylist)
-
-            # proceed with initialization that is not in the profile
+            # 5. proceed with initialization that is not in the profile
             # initialize transforms
-            if self.t_['auto_orient'] and 'flip_x' not in keylist:
+            if self.t_['auto_orient'] and 'flip_x' not in keylist2:
                 self.logger.debug(
                     "auto orient (%s)" % (self.t_['auto_orient']))
                 self.auto_orient()
 
             # initialize scale
-            if self.t_['autozoom'] != 'off' and 'scale' not in keylist:
+            if self.t_['autozoom'] != 'off' and 'scale' not in keylist2:
                 self.logger.debug("auto zoom (%s)" % (self.t_['autozoom']))
                 self.zoom_fit(no_reset=True)
 
             # initialize pan position
             # NOTE: False a possible value from historical use
             if (self.t_['autocenter'] not in ('off', False) and
-                'pan' not in keylist):
+                'pan' not in keylist2):
                 self.logger.debug(
                     "auto center (%s)" % (self.t_['autocenter']))
                 self.center_image(no_reset=True)
 
             # initialize cuts
-            if self.t_['autocuts'] != 'off' and 'cuts' not in keylist:
+            if self.t_['autocuts'] != 'off' and 'cuts' not in keylist2:
                 self.logger.debug("auto cuts (%s)" % (self.t_['autocuts']))
                 self.auto_levels()
 
-            # save the profile in the image
-            if self.use_image_profile:
-                self.checkpoint_profile()
+        # save the profile in the image
+        if self.use_image_profile:
+            self.checkpoint_profile()
 
     def apply_profile(self, profile, keylist=None):
         """Apply a profile to the viewer.
@@ -900,8 +973,17 @@ class ImageViewBase(Callback.Callbacks):
                                   callback=True)
 
     def capture_profile(self, profile):
-        self.t_.copy_settings(profile)
+        self.t_.copy_settings(profile, keylist=self.profile_keylist,
+                              callback=False)
         self.logger.debug("profile attributes set")
+
+    def capture_default_viewer_profile(self):
+        if self.default_viewer_profile is None:
+            self.default_viewer_profile = Settings.SettingGroup()
+        self.t_.copy_settings(self.default_viewer_profile,
+                              keylist=self.profile_keylist,
+                              callback=False)
+        self.logger.info("captured default profile")
 
     def checkpoint_profile(self):
         profile = self.save_profile()
@@ -918,7 +1000,7 @@ class ImageViewBase(Callback.Callbacks):
 
         """
         image = self.get_image()
-        if (image is None):
+        if image is None:
             return
 
         profile = image.get('profile', None)
@@ -935,7 +1017,6 @@ class ImageViewBase(Callback.Callbacks):
     def _update_profile_cb(self, setting, value):
         key = setting.name
         if self.use_image_profile:
-            # ? and key in self.profile_keylist
             kwargs = {key: value}
             self.save_profile(**kwargs)